Init test app binary
